+++ scripts/mpi_cmake_format
@@ -1,0 +1,46 @@
+#!/usr/bin/env python
+
+"""mpi_cmake_format
+
+License BSD-3-Clause
+Copyright (c) 2019, New York University and Max Planck Gesellschaft.
+"""
+
+# cakin install has been called and install has been sourced
+from mpi_cmake_modules import clang_format
+import argparse
+
+def _execute():
+    print("Formatting cmake files.")
+
+    ## Parser for the input arguments
+    parser = argparse.ArgumentParser(description=__doc__)
+    parser.add_argument(
+        "files_or_folders",
+        ## Store the argument in files_or_folders.
+        metavar="files_or_folders",
+        ## Define the input argument type.
+        type=str,
+        ## ...
+        nargs="+",
+        ## Help string in case the argument is wrong.
+        help="List of source files or folders.",
+    )
+
+    ## Input arguments.
+    args = parser.parse_args()
+    args.files_or_folders
+    extensions = (".cmake", "CMakeLists.txt")
+
+if __name__ == "__main__":
+    """
+    Format source files given or found recursively in the given folders.
+    """
+
+    try:
+        _execute()
+    except Exception as e:
+        import traceback
+
+        traceback.print_exc()
+        print("\nfailed with error:\n" + str(e) + "\n")
