+++ scripts/mpi_cpp_format
@@ -1,23 +1,18 @@
 #!/usr/bin/env python
 
-""" @namespace mpi_cmake_modules.clang_format
-Execute a clang-format on a desired file or folder.
+"""mpi_cpp_format
+Execute clang-format on a desired file or folder.
 
-Usage example:
-- @code rosrun mpi_cmake_module clang_format /path/to/format @endcode
-- @code rosrun mpi_cmake_module clang_format file_to_format.ext @endcode
-
-@file clang_format
-@license License BSD-3-Clause
-@copyright Copyright (c) 2019, New York University and Max Planck Gesellschaft.
+License BSD-3-Clause
+Copyright (c) 2019, New York University and Max Planck Gesellschaft.
 """
 
 # cakin install has been called and install has been sourced
 from mpi_cmake_modules import clang_format
 import argparse
 
-
 def _execute():
+    print("Formatting C/C++ files.")
 
     ## Parser for the input arguments
     parser = argparse.ArgumentParser(description=__doc__)
@@ -51,11 +46,10 @@ def _execute():
     )
 
     if not list_of_files:
-        raise RuntimeError(
-            "\nNo file to format, please indicate paths to files or directories.\n"
-        )
+        print("\nNo C/C++ file to format in the given paths.\n")
+        return
     else:
-        print("\nformatting:")
+        print("\nFormatting:")
         for f in list_of_files:
             print("\t" + str(f))
         print("")
@@ -63,7 +57,6 @@ def _execute():
     clang_format.execute_clang_format(
         clang_format_bin, clang_format_config, list_of_files
     )
-
 
 if __name__ == "__main__":
     """
